+++ NEWS.md
@@ -1,5 +1,6 @@
 # graDiEnt 1.0.1
+* Package uses "message()" instead of "print()" to communicate with user.
+* minor fix to documentation.
 
 # graDiEnt 1.0.0
-
 * First public release of package.